actually lets just bump to beta 11
not sure if theres gonna be a beta 12 or if i'm going to bump this shit into rc1 because this whole versioning thing is severly fucked.

opensb 1.3 should've been a stopgap release between 1.2 and 2.0, but it's effectively become 2.0 itself with the amount of changes thats been done. so theres a very big chance this version will end up being opensb 2.0.