initial commit of experimental branch
this will be the experimental branch for now, hopefully this wont take as long as 2.0.

this is going to be seperate from opensb-2.1 for the time being, as i intend on having that branch be for whatever trinium redesign cleber's gonna do, i don't wanna fuck his shit up, as git has quite the learning curve.

oh yeah also from now on new features should be prototyped seperately from the frontend. thats an idea ive had since 2021 but i havent been bothered to try doing so since now.